Chapter 14 Completed

+++ TweetBook/Controllers/V1/PostsController.cs
@@ -12,6 +12,7 @@ using TweetBook.Contracts.V1;
 using TweetBook.Contracts.V1.Requests;
 using TweetBook.Contracts.V1.Responses;
 using TweetBook.Domain;
+using TweetBook.Extensions;
 using TweetBook.Services;
 
 namespace TweetBook.Controllers.V1
@@ -47,6 +48,17 @@ namespace TweetBook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var ifUserOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!ifUserOwnsPost)
+            {
+                return Unauthorized(
+                    new
+                    {
+                        error = "You don't own the post"
+                    });
+            }
+
             var isPostDeleted = await _postService.DeletePostAsync(postId);
 
             if (isPostDeleted)
@@ -57,12 +69,20 @@ namespace TweetBook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest postRequest)
         {
-            var post = new Post
-            {
-                Id = postId,
-                Name = postRequest.Name
-            };
+            var ifUserOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
 
+            if (!ifUserOwnsPost)
+            {
+                return Unauthorized(
+                    new
+                    {
+                        error = "You don't own the post"
+                    });
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+            post.Name = postRequest.Name;
+                                  
             var isUpdated = await _postService.UpdatePostAsync(post);
 
             if (isUpdated)
@@ -73,7 +93,11 @@ namespace TweetBook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post
+            {
+                Name = postRequest.Name,
+                UserId = HttpContext.GetUserId()
+            };
 
             await _postService.CreatePostAsync(post);
